Added employee creation
Changed employeetype, gender og experiencelevel to string in database while it is still stored as an enum in the entity
Added helper method to take enum name as input instead of number and adding spaces with the method so e.g. "Mid Level" is stored with a space in the database.

Added form in CreateEmployee.razor for creating an employee
Added create employee to navmenu.razor
Added DTO folder for structuring inputs as a model

Added Command to create employee to be able to map the DTO into employee and address constructor

Added EmployeeRepository with methods (most of which are just placeholders)
Made CreateNewAsync method in repository which adds a employee

Removed some earlier stuff which conflicted with clean code and structuring